requirements.txt
# LangChain and LLM dependencies
langchain>=0.2.0
langchain-google-genai>=1.0.0
langchain-community>=0.2.0

# Vector database and embedding dependencies
supabase>=2.0.0
sentence-transformers>=2.2.0

# Web scraping and content processing
requests>=2.31.0
beautifulsoup4>=4.12.0
newspaper3k>=0.2.8
feedparser>=6.0.0
trafilatura>=1.6.0

# WordPress integration
python-wordpress-xmlrpc>=2.3

# Async and utilities
aiohttp>=3.8.0
python-dotenv>=1.0.0

# Web search API
google-search-results>=2.4.0

# Additional utilities
pydantic>=2.0.0
PyYAML>=6.0
pytz>=2023.0
python-dateutil>=2.8.0

# Optional: Better error handling and logging
structlog>=23.0.0
colorlog>=6.7.0

# Optional: Performance monitoring
psutil>=5.9.0 